Add feature to preparing IPL/Firmware automatically
It requires to extract and copy some files by manual to preprare
IPL/Firmware before writing it to the board so far.
This patch removes manual setup before writing IPL/Firmware.
Only thing user needs is to copy the ICUMX_loader package into tool directory,
then start building as same as so far.
After that, all required files are stored into deploy directory.

+++ build_whitebox_v4.1.sh
@@ -25,6 +25,12 @@ elif [[ "$1" != "spider" ]] && [[ "$1" != "s4sk" ]]; then
     Usage; exit
 fi
 
+ICUMX_LOADER_PATH=$(find ${BASE_DIR}/tool -name "*ICUMX_Loader_and_Flashwriter_Package_for_R-Car_S4_Starter_Kit_SDKv*.zip")
+if [ "${ICUMX_LOADER_PATH}" == "" ]; then
+  echo -e "\e[31mERROR: ICUMX loader file (zip) not found in tool directory.\e[m"
+  exit
+fi
+
 # deploy images
 cd ${BASE_DIR}
 rm -rf deploy
@@ -54,4 +60,7 @@ cp application_cpu/work/yocto/build-domd/tmp/deploy/images/${1}/tee-${1}.srec de
 cp application_cpu/work/yocto/build-domd/tmp/deploy/images/${1}/u-boot-elf-${1}.srec deploy/
 
 ls -l deploy
+
+cd tool
+./setup_ipl.sh $1
